Add files via upload
Something I found difficult was trying to figure out how to get the image to be in the middle of the screen, I tried to use 'Screen 1' in designer view to get it to be in the middle. Something easy was making the reset button since it only needed around 2 pieces of code.